Improve data flow
First attempt at creating a better data flow and implementing the
container/component pattern for decoupling the presentational components
from the redux store.
