+++ redux/store.js
@@ -17,8 +17,7 @@ const bindMiddleware = (middleware) => {
 // Define the initial state of the store .
 const initialState = {
   date_filter: { from_: moment().startOf('year'), to: moment().endOf('year') },
-  fatalities: [],
-  view: 'apdView'
+  fatalities: []
 }
 
 // Define the redux action types.
@@ -26,7 +25,6 @@ export const actionTypes = {
   FETCH_DATA: 'FETCH_DATA',
   FETCH_DATA_ASYNC: 'FETCH_DATA_ASYNC',
   SELECT_DATE: 'SELECT_DATE',
-  SETVIEW: 'SETVIEW'
 }
 
 // Define reducers.
@@ -34,7 +32,6 @@ export const reducer = (state = initialState, action) => {
   switch (action.type) {
     case actionTypes.FETCH_DATA:
     case actionTypes.SELECT_DATE:
-    case actionTypes.SETVIEW:
       return {
         ...state, ...action.payload
       }
@@ -58,11 +55,6 @@ export const fetchDataAsync = (date_filter) => ({
 export const selectDate = (date_filter) => ({
   type: actionTypes.SELECT_DATE,
   payload: { date_filter }
-});
-
-export const setView = (view) => ({
-  type: actionTypes.SETVIEW,
-  payload: { view }
 });
 
 // Initialize store.
